+++ example/shell/set_var_now.dart
@@ -3,10 +3,8 @@ import 'common.dart';
 Future<void> main() async {
   var now = DateTime.now().toUtc().toIso8601String();
   await shell.run('''
-# Change the env file location
-ds env var set TEST_VAR_NOW $now
-
-ds env var dump
-
+ds -v env var get TEST_VAR_NOW  
+ds -v env var set TEST_VAR_NOW $now
+ds -v env var get TEST_VAR_NOW
 ''');
 }