+++ lib/src/bin/shell/env_alias_delete.dart
@@ -5,47 +5,54 @@ import 'package:process_run/shell.dart';
 import 'package:process_run/src/bin/shell/env.dart';
 import 'package:process_run/src/common/import.dart';
 
-class ShellEnvAliasSetCommand extends ShellEnvCommandBase {
-  ShellEnvAliasSetCommand()
+class ShellEnvAliasDeleteCommand extends ShellEnvCommandBase {
+  ShellEnvAliasDeleteCommand()
       : super(
-          name: 'set',
-          description: 'Set process_run alias',
+          name: 'delete',
+          description: 'Delete an alias from a user/local config file',
         );
 
   @override
   void printUsage() {
-    stdout.writeln('ds env alias set <name> <command with space>');
+    stdout.writeln('ds env alias delete <name> [<name2>...]');
     super.printUsage();
   }
 
   @override
   FutureOr<bool> onRun() async {
     var rest = results.rest;
-    if (rest.length < 2) {
-      stderr.writeln('At least 2 arguments expected');
+    if (rest.isEmpty) {
+      stderr.writeln('At least 1 arguments expected');
       exit(1);
     } else {
       if (verbose) {
         stdout.writeln('file $label: $envFilePath');
         stdout.writeln('before: ${jsonEncode(ShellEnvironment().aliases)}');
       }
-      var alias = rest[0];
-      var command = rest.sublist(1).join(' ');
+
       var fileContent = await envFileReadOrCreate();
-      if (fileContent.addAlias(alias, command)) {
+      var modified = false;
+      for (var name in rest) {
+        modified = fileContent.deleteAlias(name) ?? modified;
+      }
+      if (modified) {
+        if (verbose) {
+          stdout.writeln('writing file');
+        }
         await fileContent.write();
       }
+
       // Force reload
       shellEnvironment = null;
       if (verbose) {
-        stdout.writeln('After: ${jsonEncode(ShellEnvironment().aliases)}');
+        stdout.writeln('After: ${jsonEncode(ShellEnvironment().vars)}');
       }
       return true;
     }
   }
 }
 
-/// Direct shell env Alias Set run helper for testing.
+/// Direct shell env Var Set run helper for testing.
 Future<void> main(List<String> arguments) async {
-  await ShellEnvAliasSetCommand().parseAndRun(arguments);
+  await ShellEnvAliasDeleteCommand().parseAndRun(arguments);
 }