+++ lib/src/bin/shell/env_edit.dart
@@ -1,7 +1,6 @@
 import 'dart:io';
 
 import 'package:process_run/shell.dart';
-import 'package:process_run/src/bin/shell/shell.dart';
 import 'package:process_run/src/common/import.dart';
 
 import 'env.dart';