+++ lib/src/bin/shell/env_path_delete.dart
@@ -5,16 +5,16 @@ import 'package:process_run/shell.dart';
 import 'package:process_run/src/bin/shell/env.dart';
 import 'package:process_run/src/common/import.dart';
 
-class ShellEnvPathPrependCommand extends ShellEnvCommandBase {
-  ShellEnvPathPrependCommand()
+class ShellEnvPathDeleteCommand extends ShellEnvCommandBase {
+  ShellEnvPathDeleteCommand()
       : super(
-          name: 'prepend',
-          description: 'Prepend path for executable lookup',
+          name: 'delete',
+          description: 'Delete paths for executable lookup in config file',
         );
 
   @override
   void printUsage() {
-    stdout.writeln('ds env path prepend <path1> <path2>...');
+    stdout.writeln('ds env path delete <path1> <path2>...');
     super.printUsage();
   }
 
@@ -26,10 +26,11 @@ class ShellEnvPathPrependCommand extends ShellEnvCommandBase {
       exit(1);
     } else {
       if (verbose) {
+        stdout.writeln('File $label: $envFilePath');
         stdout.writeln('before: ${jsonEncode(ShellEnvironment().paths)}');
       }
       var fileContent = await envFileReadOrCreate();
-      if (fileContent.prependPaths(paths)) {
+      if (fileContent.deletePaths(paths)) {
         await fileContent.write();
       }
       // Force reload
@@ -44,5 +45,5 @@ class ShellEnvPathPrependCommand extends ShellEnvCommandBase {
 
 /// Direct shell env Var Set run helper for testing.
 Future<void> main(List<String> arguments) async {
-  await ShellEnvPathPrependCommand().parseAndRun(arguments);
+  await ShellEnvPathDeleteCommand().parseAndRun(arguments);
 }